requirements.txt
# requirements.txt

Pillow>=11.1.0,<12.0.0
transformers>=4.48.0,<5.0.0
torch>=2.5.1,<3.0.0
pytest~=7.4.4
setuptools~=75.1.0
tqdm~=4.66.5
cairosvg~=2.7.1